requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
cbor==1.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
datasets==3.5.0
debugpy==1.8.14
decorator==5.2.1
dill==0.3.8
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.18.0
FlagEmbedding==1.3.4
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2024.12.0
huggingface-hub==0.30.2
idna==3.10
ijson==3.3.0
inscriptis==2.6.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
ir_datasets==0.5.10
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
kiwisolver==1.4.8
lxml==5.3.2
lz4==4.4.4
mabwiser==2.7.4
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.15.2
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scikit-surprise==1.1.4
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==78.1.0
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
trec-car-tools==2.6
typing_extensions==4.13.2
tzdata==2025.2
unlzw3==0.2.3
urllib3==2.4.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wcwidth==0.2.13
widgetsnbextension==4.0.14
xxhash==3.5.0
yarl==1.20.0
zlib-state==0.1.9